Revert "Moved things out of the API, in order to use extensions"
This reverts commit 4b42ac9ae0562d22c2ad38c117013c146038796d.
