+++ src/api/java/com/machina/api/planet/attribute/PlanetAttributeList.java
@@ -91,7 +91,7 @@ public class PlanetAttributeList extends NBTList<PlanetAttribute<?>, CompoundNBT
 		}
 	}
 	
-	public static UnaryOperator<Integer> intOperator(Operation operation, Integer value) {
+	public UnaryOperator<Integer> intOperator(Operation operation, Integer value) {
 		switch (operation) {
 		case ADDITION: return old -> value + old;
 		case SUBSTRACTION: return old -> old - value;
