+++ src/api/java/com/machina/api/planet/attribute/PlanetAttributeType.java
@@ -30,13 +30,17 @@
 
 package com.machina.api.planet.attribute;
 
+import static com.machina.api.ModIDs.MACHINA;
+
 import java.util.Random;
 import java.util.function.Function;
 
-import net.minecraft.nbt.INBT;
+import com.matyrobbrt.lib.registry.annotation.RegistryHolder;
 
+import net.minecraft.nbt.INBT;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 
+@RegistryHolder(modid = MACHINA)
 public class PlanetAttributeType<T> extends ForgeRegistryEntry<PlanetAttributeType<?>> {
 
 	private final String measureUnit;
