--- src/api/java/com/machina/api/starchart/package-info.java
@@ -1,1 +1,0 @@
-package com.machina.api.starchart;
