+++ src/main/java/com/machina/events/ForgeEvents.java
@@ -30,14 +30,11 @@
 
 package com.machina.events;
 
-import static com.machina.api.ModIDs.MACHINA;
-
 import com.machina.Machina;
-import com.machina.starchart.StarchartImpl;
-import com.machina.world.DynamicDimensionHelper;
+import com.machina.api.world.DynamicDimensionHelper;
+import com.machina.api.world.data.StarchartData;
 
 import net.minecraft.entity.player.ServerPlayerEntity;
-
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.entity.item.ItemTossEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent.PlayerLoggedInEvent;
@@ -45,7 +42,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
 
-@Mod.EventBusSubscriber(modid = MACHINA, bus = Bus.FORGE)
+@Mod.EventBusSubscriber(modid = Machina.MOD_ID, bus = Bus.FORGE)
 public class ForgeEvents {
 
 	@SubscribeEvent
@@ -58,7 +55,7 @@ public class ForgeEvents {
 		if (event.getEntity().level.isClientSide) {
 			return;
 		}
-		StarchartImpl.getDefaultInstance(event.getPlayer().getServer()).debugStarchart();
+		StarchartData.getDefaultInstance(event.getPlayer().getServer()).debugStarchart();
 		DynamicDimensionHelper.sendPlayerToDimension((ServerPlayerEntity) event.getPlayer(),
 				DynamicDimensionHelper.createPlanet(event.getPlayer().getServer(), String.valueOf(0)),
 				event.getPlayer().position());
