+++ src/main/java/com/machina/init/KeyBindingsInit.java
@@ -30,9 +30,9 @@
 
 package com.machina.init;
 
-import static com.machina.api.ModIDs.MACHINA;
-
 import java.awt.event.KeyEvent;
+
+import com.machina.Machina;
 
 import net.minecraft.client.settings.KeyBinding;
 
@@ -45,7 +45,7 @@ import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
 @OnlyIn(Dist.CLIENT)
-@Mod.EventBusSubscriber(modid = MACHINA, bus = Bus.MOD, value = Dist.CLIENT)
+@Mod.EventBusSubscriber(modid = Machina.MOD_ID, bus = Bus.MOD, value = Dist.CLIENT)
 public final class KeyBindingsInit {
 
 	public static final KeyBinding DEV_PLANET_CREATION_SCREEN = create("dev_planet_creation_screen", KeyEvent.VK_C);
@@ -58,7 +58,7 @@ public final class KeyBindingsInit {
 	}
 
 	private static KeyBinding create(String name, int key) {
-		return new KeyBinding("key." + MACHINA + "." + name, key, "key.category." + MACHINA);
+		return new KeyBinding("key." + Machina.MOD_ID + "." + name, key, "key.category." + Machina.MOD_ID);
 	}
 
 	public static boolean isKeyDown(KeyBinding key) {
