--- src/main/java/com/machina/starchart/package-info.java
@@ -1,1 +1,0 @@
-package com.machina.starchart;
